Improve tracking of the channels
Use the index of the channel instead of matching
on the full name of the channel.

This prevents a bug in such a rare scenario:
if user has more than 10 channels and has asked
to listen to the channel #1, he will also listen
to channels #11, #12, etc due to substring mathing.